Vue/homework/07-01-vue-with-drf/vue_ws_8_5/back/requirements.txt
asgiref==3.7.2
Django==4.2.4
django-cors-headers==4.2.0
django-seed==0.3.1
djangorestframework==3.14.0
Faker==25.1.0
python-dateutil==2.9.0.post0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
toposort==1.10
typing_extensions==4.7.1
tzdata==2023.3
